+++ jrnl/messages/__init__.py
@@ -1,0 +1,7 @@
+from .Message import Message
+from .MsgStyle import MsgStyle
+from .MsgText import MsgText
+
+Message = Message
+MsgStyle = MsgStyle
+MsgText = MsgText
